Fix variable export in Makefile
Makefile runs each line of a target in a separate shell, so things like
`export` and `source` do not propagate to the next lines, you must run
the commands in a single shell, which is what I've done in this commit.

With out this change, `make build` fails on my debian linux box.